Remove unportable and unnecessary time zone fix up code.
The code using setenv("TZ") doesn't compile when using Microsoft Visual C++
compiler as it doesn't have this (POSIX) function. And if it compiled, it
still wouldn't work as changing TZ environment variable doesn't work as
expected by this code under Microsoft Windows. And it seems to be totally
unnecessary anyhow, as we are working with the broken down time representation
which is already in UTC and not time_t anyhow.

So just remove it, not doing this doesn't change struct tm contents at all.