Add line number to parse error messages.
Just saying "Failed to parse" was singularly unhelpful, at least mention at
which line number the error has happened. Perhaps column number reporting
could be added in the future, but the exact position of the error is not
always well-defined, so don't do it for now.

Replace all occurrences of "throw toml_parse_exception{...}" with the calls to
throw_parse_exception(...), this will make it easier to customize the error
reporting later if necessary by updating just this function.