Avoid clang warnings in cpptoml sources.
Use noreturn attribute for clang as well, otherwise every use of
throw_parse_exception() results in a -Wno-return warning.

The analyzer_noreturn attribute was inappropriate here as it indicates that
the function can, actually, return, but should just be considered as not
returning for static analysis purposes, which was not the case here. And it
doesn't help with compilation warnings at all, so all it did was to prevent
the noreturn attribute from being used.