Rename cpptomlConfig.cmake to cpptomlConfig.cmake.in.
This file was being erroneously found for some reason when cpptoml is
used via the user package repository in CMake, but only on some
systems. I don't know what the difference is, but having this be a .in
fixes the issue and causes CMake to find the correct
cpptomlConfig.cmake.